Split out crawl delay cache
DOIDownloader class no longer has a refernce to a cache file. Instead we give it a dict-like object, where it can get and set domain delays.

We have an example implementation of such an object in __main__.py, which replicates the old behaviour.
We could add a similar class that stores crawl delays in the sqlite database.

Fixes #26